requirements.txt
Jinja2
python-dotenv
lxml
pre-commit